pyproject.toml
[project]
name = "ayon-server"
version = "1.7.3"
description = ""
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
  "aiocache >=0.12.2",
  "aiofiles >=23.2.1",
  "aioshutil >=1.3",
  "asyncpg >=0.29.0",
  "boto3 >=1.35.22",
  "codenamize >=1.2.3",
  "cryptography >=44.0.0",
  "email-validator >=1.1.3",
  "fastapi >=0.115",
  "geoip2 >=4.8.0",
  "granian >=1.1.1",
  "gunicorn >=23.0",
  "httpx >=0.26",
  "loguru>=0.7.3",
  "orjson >=3.8.6",
  "pillow >=10.3.0",
  "psutil >=5.9.4",
  "pydantic ==1.10.18", # 1.10.18 is the last 1.x version
  "pyjwt ==2.9.0", # Do not upgrade before migrating to string only subs
  "pyyaml >=6.0.1",
  "redis >=5.0.1",
  "rich >=13.4.2",
  "semver >=3.0.1",
  "shortuuid >=1.0.11",
  "strawberry-graphql[fastapi] >=0.243.0",
  "toml >=0.10.2",
  "typer >=0.12.5",
  "unidecode>=1.3.8",
  "user-agents >=2.2.0",
  "uvicorn[standard] >=0.25",
]

[dependency-groups]
dev = [
  "asyncpg-stubs >=0.29.1",
  "mypy >=1.11",
  "pre-commit >=2.20.0",
  "pytest >=7.0.0",
  "pytest-order >=1.0.1",
  "ruff >=0.6.9",
  "types-aiofiles >=23.2.0.20240403",
  "types-psutil >=5.9.5.20240516",
  "types-pyyaml >=6.0.12.20240311",
  "types-toml >=0.10.8",
  "types-boto3 >=1.0.2",
]

[tool.ruff]
fix = true
line-length = 88

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"


[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

ignore = [
    "ISC001", # incompatible with formatting
    "E721",  # do not compare types, i'll fix this later :)
    "C901",  # too complex. C'mon - this is a complex project
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = [
  "pydantic.mypy"
]

check_untyped_defs = true
disallow_any_generics = true
exclude = "tests/|venv/"
explicit_package_bases = true
no_implicit_reexport = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
# follow_imports = "silent"
# strict = true
# disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
  "user_agents",
  "tomllib",
  "codenamize",
  "aiocache"
]
ignore_missing_imports = true
